complete 4.21 (DELETE for blog creator only & tests)

+++ osa3/blogilista/contollers/blogs.js
@@ -2,6 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const blog = require('../models/blog')
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -38,8 +39,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (decodedToken.id.toString() === blogToDelete.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({
+      error: 'cannot delete blog created by different user'
+    })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
